llm/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
eval_type_backport==0.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jsonpath-python==1.0.6
mistralai==1.5.0
mypy-extensions==1.0.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sniffio==1.3.1
typing-inspect==0.9.0
typing_extensions==4.12.2
